# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AgentHub AI - Multi-tenant SaaS API for AI-powered messaging agents
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workspaces/{workspaceId}/agents:
    get:
      operationId: listAgents
      tags: [agents]
      summary: List all AI agents in a workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agent"
    post:
      operationId: createAgent
      tags: [agents]
      summary: Create a new AI agent in a workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentInput"
      responses:
        "201":
          description: Created agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
  /agents/{agentId}:
    get:
      operationId: getAgent
      tags: [agents]
      summary: Get an AI agent by ID
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "404":
          description: Not found
    patch:
      operationId: updateAgent
      tags: [agents]
      summary: Update an AI agent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentUpdate"
      responses:
        "200":
          description: Updated agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "404":
          description: Not found
    delete:
      operationId: deleteAgent
      tags: [agents]
      summary: Delete an AI agent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /agents/{agentId}/toggle-status:
    patch:
      operationId: toggleAgentStatus
      tags: [agents]
      summary: Toggle an agent's active/inactive status
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "404":
          description: Not found

  # ─── API Keys ─────────────────────────────────────────────────────────────
  /workspaces/{workspaceId}/api-keys:
    get:
      operationId: listApiKeys
      tags: [api-keys]
      summary: List API keys for a workspace (masked)
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of masked API keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiKey"
    post:
      operationId: createApiKey
      tags: [api-keys]
      summary: Add an API key to a workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiKeyInput"
      responses:
        "201":
          description: Created API key (masked)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKey"
  /api-keys/{keyId}:
    delete:
      operationId: deleteApiKey
      tags: [api-keys]
      summary: Delete an API key
      parameters:
        - name: keyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /api-keys/{keyId}/test:
    post:
      operationId: testApiKey
      tags: [api-keys]
      summary: Test if an API key is valid
      parameters:
        - name: keyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyTestResult"

  # ─── AI Brains ────────────────────────────────────────────────────────────
...
  /agents/{agentId}/chat:
    post:
      operationId: chatWithAgent
      tags: [agents]
      summary: Send a test message to an agent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatInput"
      responses:
        "200":
          description: Agent reply
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatReply"

  # ─── Dashboard ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Agent:
      type: object
      required: [id, workspaceId, name, role, status, provider, model, temperature, maxTokens, emojiLevel, humorLevel, creativity, createdAt, updatedAt]
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        businessId:
          type: ["integer", "null"]
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
        job:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        language:
          type: string
        tone:
          type: string
          enum: [professional, friendly, casual, formal, empathetic]
        personality:
          type: ["string", "null"]
        instructions:
          type: ["string", "null"]
        goal:
          type: ["string", "null"]
        emojiLevel:
          type: integer
          minimum: 0
          maximum: 10
        humorLevel:
          type: integer
          minimum: 0
          maximum: 10
        creativity:
          type: integer
          minimum: 0
          maximum: 10
        temperature:
          type: number
          minimum: 0
          maximum: 2
        maxTokens:
          type: integer
        provider:
          type: string
          enum: [deepseek, gemini, openai, claude]
        model:
          type: string
        status:
          type: string
          enum: [active, inactive, draft]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AgentInput:
      type: object
      required: [name, role, provider, model]
      properties:
        businessId:
          type: integer
        name:
          type: string
          minLength: 1
        role:
          type: string
          minLength: 1
        job:
          type: string
        description:
          type: string
        language:
          type: string
        tone:
          type: string
          enum: [professional, friendly, casual, formal, empathetic]
        personality:
          type: string
        instructions:
          type: string
        goal:
          type: string
        emojiLevel:
          type: integer
          minimum: 0
          maximum: 10
        humorLevel:
          type: integer
          minimum: 0
          maximum: 10
        creativity:
          type: integer
          minimum: 0
          maximum: 10
        temperature:
          type: number
          minimum: 0
          maximum: 2
        maxTokens:
          type: integer
        provider:
          type: string
          enum: [deepseek, gemini, openai, claude]
        model:
          type: string
    AgentUpdate:
      type: object
      properties:
        businessId:
          type: integer
        name:
          type: string
          minLength: 1
        role:
          type: string
        job:
          type: string
        description:
          type: string
        language:
          type: string
        tone:
          type: string
          enum: [professional, friendly, casual, formal, empathetic]
        personality:
          type: string
        instructions:
          type: string
        goal:
          type: string
        emojiLevel:
          type: integer
          minimum: 0
          maximum: 10
        humorLevel:
          type: integer
          minimum: 0
          maximum: 10
        creativity:
          type: integer
          minimum: 0
          maximum: 10
        temperature:
          type: number
          minimum: 0
          maximum: 2
        maxTokens:
          type: integer
        provider:
          type: string
          enum: [deepseek, gemini, openai, claude]
        model:
          type: string
        status:
          type: string
          enum: [active, inactive, draft]

    # ─── Dashboard ───────────────────────────────────────────────────────────
...
    ApiKey:
      type: object
      required: [id, workspaceId, provider, label, keyPreview, createdAt]
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        provider:
          type: string
          enum: [deepseek, gemini, openai, claude]
        label:
          type: string
        keyPreview:
          type: string
        createdAt:
          type: string
          format: date-time
    ApiKeyInput:
      type: object
      required: [provider, label, key]
      properties:
        provider:
          type: string
          enum: [deepseek, gemini, openai, claude]
        label:
          type: string
          minLength: 1
        key:
          type: string
          minLength: 8
    ApiKeyTestResult:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
        message:
          type: string

    # ─── AI Brain ─────────────────────────────────────────────────────────────
...
    ChatInput:
      type: object
      required: [message]
      properties:
        message:
          type: string
          minLength: 1
        conversationHistory:
          type: array
          items:
            type: object
            required: [role, content]
            properties:
              role:
                type: string
                enum: [user, assistant]
              content:
                type: string
    ChatReply:
      type: object
      required: [reply, provider, model]
      properties:
        reply:
          type: string
        provider:
          type: string
        model:
          type: string
        tokensUsed:
          type: integer