# lib/api-zod/src/generated/types/apiKey.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AgentHub AI - Multi-tenant SaaS API for AI-powered messaging agents
 * OpenAPI spec version: 0.1.0
 */
import type { ApiKeyProvider } from './apiKeyProvider';

export interface ApiKey {
  id: number;
  workspaceId: number;
  provider: ApiKeyProvider;
  label: string;
  keyPreview: string;
  createdAt: Date;
}