# lib/api-zod/src/generated/types/apiKeyInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AgentHub AI - Multi-tenant SaaS API for AI-powered messaging agents
 * OpenAPI spec version: 0.1.0
 */
import type { ApiKeyInputProvider } from './apiKeyInputProvider';

export interface ApiKeyInput {
  provider: ApiKeyInputProvider;
  /** @minLength 1 */
  label: string;
  /** @minLength 8 */
  key: string;
}